ops/solid_pass/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
tabulate>=0.9.0
sseclient-py>=1.8.0
psycopg2-binary>=2.9.9

beautifulsoup4>=4.12.0
